Allow to compile nativelib in Scala 3
